==> analysis/02_fangraphs_datacleaning.ipynb <==
# Notebook 2: Cleaning Fangraphs Data

### Introduction

This notebook is intended as a reference for cleaning data scraped from Fangraphs. There are no special libraries required, as most of the cleaning can be done with pandas.

import pandas as pd

### Cleaning Function

All of the major data cleaning can be done in one cleaning function. There are some jokers at Fangraphs who entered in projections for "Rob W00t3n", so I hard-coded a fix, but aside from that, it's mostly replacing characters and generating an 'ID' variable based on the player name so it's possible to combine this with the Baseball Reference data.

def clean_fangraphs_season(csvpath, suffix):
    file = pd.read_csv(csvpath)
    if csvpath == 'projected_2017.csv':
        file.iloc[665, 0] = 'Rob Wooten'
        file.iloc[826, 0] = 'Rob Wooten'
    file['ID'] = file.Player.str.replace("’", '').str.replace('.', '').str.replace('-', '')
    file['ID'] = file.ID.str.extract(r'\s(\b[a-zA-Z]{1,5})[A-z]*$') + file.ID.str.extract(r'(^[A-z]{2})')
    file['ID'] = file.ID.str.lower()
    file['ID'] = file.ID.astype(str) + '01'
    cols = file.columns.tolist()
    new_cols = [cols[-1]] + [i for i in cols[0:-1]]
    file_v2 = file[new_cols]
    new_names = [(i,i+ suffix) for i in file_v2.iloc[:, 2:].columns.values]
    file_v2.rename(columns = dict(new_names), inplace=True)
    return file_v2
    

### Saving to CSV

I'll call the function on my scraped data, and save the resulting pandas dataframe as a csv.

projected_2017_c = clean_fangraphs_season('../data/projected_2017.csv', '_2017')
projected_2018_c = clean_fangraphs_season('../data/projected_2018.csv', '_2018')
projected_2019_c = clean_fangraphs_season('../data/projected_2019.csv', '_2019')


# projected_2017_c.to_csv('../data/projected_2017_c.csv', index = False)
# projected_2018_c.to_csv('../data/projected_2018_c.csv', index = False)
# projected_2019_c.to_csv('../data/projected_2019_c.csv', index = False)
