==> analysis/03_fangraphs_bbref_combine.ipynb <==
# Notebook 3: Combining Fangraphs and Baseball Reference Data

### Introduction

From the previous two notebooks, we now have clean projection and season data to use for modeling. In this notebook, I'll combine the relevant datasets to form the training, validation, and test set.

import pandas as pd

### Data for Training Model

My training set will use season data from 2016 and projection data from 2017 to predict quality starts in 2017. This code goes over how I combine them all, and what rows/players were dropped.

 Need to have:
- season QS data for 2017
- projection (all) data for 2017
- season (all) data for 2016

qs_17 = pd.read_csv('../data/season_2017_c.csv')
qs_17_simple = qs_17[['Full_Name', 'ID', 'QS_2017']]

season_2016 = pd.read_csv('../data/season_2016_c.csv')

This dataset (s16_qs17) has a QS_2017 value for all rows. However, not all rows will have 2016 season data. This is all taken from baseball reference.

s16_qs17 = season_2016.merge(qs_17_simple, left_on =['ID', 'Full_Name'], 
                                    right_on = ['ID', 'Full_Name'], how = 'right')
# Performing a right join here means that all rows have a QS_2017 value, and not all rows have 2016 season data.

s16_qs17.shape

Now we will read in the projection data and perform the first match, on player ID. Note that player ID has been constructed in the Fangraphs projection data based on the Player Name. It will not necessarily match the ID given by Baseball Reference, and will need to be updated for non-matches.

proj17 = pd.read_csv('../data/projected_2017_c.csv')

train_matchID = s16_qs17.merge(proj17, left_on =['ID'], 
                                    right_on = ['ID'], how = 'inner')


This dataset is the result of the first match between the 2017QS/2016 season data and the 2017 projection data.

train_matchID

Let's capture the rest of the unmatched data, where there was no ID that lined up.

train_matchID_unmatched_bbref = s16_qs17[~s16_qs17.ID.isin(train_matchID.ID)]
train_matchID_unmatched_proj  = proj17[~proj17.ID.isin(train_matchID.ID)].drop(['ID'], axis = 1)

Now let's perform the second match, on Full_Name. 

train_match_Full_Name = train_matchID_unmatched_bbref.merge(train_matchID_unmatched_proj, left_on =['Full_Name'], 
                                    right_on = ['Player'], how = 'inner')

train_match_Full_Name

train_matchID_unmatched_bbref_r2 = s16_qs17[~s16_qs17.ID.isin(train_matchID.ID) & ~s16_qs17.ID.isin(train_match_Full_Name.ID)]
train_matchID_unmatched_proj_r2  = proj17[~proj17.ID.isin(train_matchID.ID) & ~proj17.ID.isin(train_match_Full_Name.ID)]

#Can confirm, none of these players appeared in the 2017 projection data.
train_matchID_unmatched_bbref_r2

Now that I've found all possible player matches, I'll finalize the training data, and output it as a csv.

training_data_model = pd.concat([train_matchID, train_match_Full_Name], ignore_index = True, axis = 0)
#training_data_model.to_csv('training_data_model.csv', index = False)

### Data for Validation Model

My validation set will use season data from 2017 and projection data from 2018 to predict quality starts in 2018. This code goes over how I combine them all, and what rows/players were dropped. This is similar to what I did for the training data, but for a different year.

 Need to have:
- season QS data for 2018
- projection (all) data for 2018
- season (all) data for 2017

qs_18 = pd.read_csv('../data/season_2018_c.csv')
qs_18_simple = qs_18[['Full_Name', 'ID', 'QS_2018']]

season_2017 = pd.read_csv('../data/season_2017_c.csv')
s17_qs18 = season_2017.merge(qs_18_simple, left_on =['ID', 'Full_Name'], 
                                    right_on = ['ID', 'Full_Name'], how = 'right')

proj18 = pd.read_csv('../data/projected_2018_c.csv')

train_matchID = s17_qs18.merge(proj18, left_on =['ID'], 
                                    right_on = ['ID'], how = 'inner')

train_matchID_unmatched_bbref = s17_qs18[~s17_qs18.ID.isin(train_matchID.ID)]
train_matchID_unmatched_proj  = proj18[~proj18.ID.isin(train_matchID.ID)].drop(['ID'], axis = 1)

train_match_Full_Name = train_matchID_unmatched_bbref.merge(train_matchID_unmatched_proj, left_on =['Full_Name'], 
                                    right_on = ['Player'], how = 'inner')

train_matchID_unmatched_bbref_r2 = s17_qs18[~s17_qs18.ID.isin(train_matchID.ID) & ~s17_qs18.ID.isin(train_match_Full_Name.ID)]
train_matchID_unmatched_proj_r2  = proj18[~proj18.ID.isin(train_matchID.ID) & ~proj18.ID.isin(train_match_Full_Name.ID)]

train_matchID_unmatched_bbref_r2

validation_data_model = pd.concat([train_matchID, train_match_Full_Name], ignore_index = True, axis = 0)
#validation_data_model.to_csv('validation_data_model.csv', index = False)

### Data for Test Model

My test set will use season data from 2018 and projection data from 2019 to predict quality starts in 2019. This code goes over how I combine them all, and what rows/players were dropped. This is similar to what I did for the training and validation data, but for a different year.
Need to have:

- season QS data for 2019
- projection (all) data for 2019
- season (all) data for 2018

qs_19 = pd.read_csv('../data/season_2019_c.csv')
qs_19_simple = qs_19[['Full_Name', 'ID', 'QS_2019']]

season_2018 = pd.read_csv('../data/season_2018_c.csv')


s18_qs19 = season_2018.merge(qs_19_simple, left_on =['ID', 'Full_Name'], 
                                    right_on = ['ID', 'Full_Name'], how = 'right')

proj19 = pd.read_csv('../data/projected_2019_c.csv')

train_matchID = s18_qs19.merge(proj19, left_on =['ID'], 
                                    right_on = ['ID'], how = 'inner')

train_matchID_unmatched_bbref = s18_qs19[~s18_qs19.ID.isin(train_matchID.ID)]
train_matchID_unmatched_proj  = proj19[~proj19.ID.isin(train_matchID.ID)].drop(['ID'], axis = 1)

train_match_Full_Name = train_matchID_unmatched_bbref.merge(train_matchID_unmatched_proj, left_on =['Full_Name'], 
                                    right_on = ['Player'], how = 'inner')

train_matchID_unmatched_bbref_r2 = s18_qs19[~s18_qs19.ID.isin(train_matchID.ID) & ~s18_qs19.ID.isin(train_match_Full_Name.ID)]
train_matchID_unmatched_proj_r2  = proj19[~proj19.ID.isin(train_matchID.ID) & ~proj19.ID.isin(train_match_Full_Name.ID)]

train_matchID_unmatched_bbref_r2

test_data_model = pd.concat([train_matchID, train_match_Full_Name], ignore_index = True, axis = 0)
#test_data_model.to_csv('test_data_model.csv', index = False)

test_data_model.shape
